fixed bugs with float.types of data

+++ db/migrate/20160923142649_create_stata_fuels.rb
@@ -2,13 +2,13 @@ class CreateStataFuels < ActiveRecord::Migration[5.0]
   def change
     create_table :stata_fuels do |t|
       
-      t.float	        :odo 			          # текущие показания одометра, км
-      t.float	        :odo_delta		      # изменения показателей одометра (пробег), км
-      t.float	        :refueling		      # кол-во заправленного топлива, л
-      t.float	        :price_fuel		      # стоимость заправленого топлива, грн
-      t.string	      :fuel_station	      # заправка
-      t.string	      :fuel_type		      # марка бензина
-      t.references    :bike #, foreign_key: true
+      t.numeric	        :odo 			          # текущие показания одометра, км
+      t.numeric	        :odo_delta		      # изменения показателей одометра (пробег), км
+      t.numeric	        :refueling		      # кол-во заправленного топлива, л
+      t.numeric	        :price_fuel		      # стоимость заправленого топлива, грн
+      t.string  	      :fuel_station	      # заправка
+      t.string	        :fuel_type		      # марка бензина
+      t.references      :bike #, foreign_key: true
 
       t.timestamps
     end
